Scale the DrawingCanvas with the screen size
